+++ packages/blockchain-wallet-v4-frontend/src/scenes/TaxCenter/index.tsx
@@ -1,32 +1,16 @@
 import React, { useEffect, useState } from 'react'
 import { connect } from 'react-redux'
-import moment from 'moment'
 import { bindActionCreators } from 'redux'
 
 import { actions, selectors } from 'data'
 
 import TaxCenter from './template'
-
-const FIRST_YEAR = 2014
-
-const getLimits = (option) => {
-  if (option === 0) {
-    return {
-      from: moment().year(FIRST_YEAR).startOf('year').toISOString(),
-      to: moment().subtract(1, 'days').toISOString()
-    }
-  }
-
-  return {
-    from: moment().year(option).startOf('year').toISOString(),
-    to: moment().year(option).endOf('year').toISOString()
-  }
-}
+import { getLimits } from './utils'
 
 const TaxCenterContainer = ({ addresses, modalActions, options, reports, taxCenterActions }) => {
   const [option, setOption] = useState(0)
 
-  const onGenerateReportClick = () => {
+  const handleClick = () => {
     const limits = getLimits(option)
 
     taxCenterActions.createReport({ ...addresses, ...limits })
@@ -42,10 +26,10 @@ const TaxCenterContainer = ({ addresses, modalActions, options, reports, taxCent
   return (
     <TaxCenter
       value={option}
-      handleChange={handleChange}
+      onChange={handleChange}
       options={options}
       reports={reports}
-      onGenerateReportClick={onGenerateReportClick}
+      onClick={handleClick}
     />
   )
 }
